Feat: Resolution fully working and syncing, moved files for more organization
