--- MacVE/Playback/Bridging/VideoCompositor.swift
@@ -1,3 +1,0 @@
-//Created by Lugalu on 13/06/24.
-
-import AVKit
